requirements.txt
langchain
python-dotenv
langchain-core
langchain-community
bs4
faiss-cpu
google-genai>=0.1.0
langchain[google-genai]
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain-groq
langgraph
langgraph-cli[inmem]